# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Local Google Classroom Dashboard", "version": "1.0.0" },
  "paths": {
...
    "/api/sync": {
      "post": {
        "summary": "Run Sync",
        "description": "Queue (hosted) or run (desktop) a sync of the calling user's cache.\n\n§12: /api/sync must never touch another user's data. Desktop: the local\nowner (token.json) — sync runs inline as before. Hosted: the session\nuser's oauth_tokens — the request only flags ``sync_requested`` and\nanswers ``{\"ok\": True, \"queued\": True, \"status\": \"queued\"}`` immediately;\nthe worker container performs the actual Classroom fan-out (DDoS plan\n§9: never run the full sync inside the HTTP request).\n\nHosted conflict mapping: a sync already in flight for THIS user → 409;\na manual request inside the per-user cooldown → 429 + Retry-After; the\nbackground scheduler bypasses the cooldown. Other users sync\nindependently. Rate-limit buckets (§39) stay in middleware.\n\n``restart=true`` (ADR-0032) is the dashboard's answer to \"this sync is\nstuck\". It is the ONE case where an in-flight sync does not answer 409: a\nclaim older than ``SYNC_STUCK_SECONDS`` is released and a new one queued.\nA younger claim still answers 409 — a long but progressing Classroom\nimport must never be interrupted, and the cooldown is skipped only because\nthe abandoned claim is by definition older than it.",
        "operationId": "run_sync_api_sync_post",
        "parameters": [
          {
            "name": "restart",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Abandon a stuck sync and start a new one (ADR-0032).",
              "default": false,
              "title": "Restart"
            },
            "description": "Abandon a stuck sync and start a new one (ADR-0032)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SyncResult" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/cache": {
      "delete": {
        "summary": "Clear Cache",
        "operationId": "clear_cache_api_cache_delete",
        "parameters": [
          {
            "name": "confirm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Confirm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Clear Cache Api Cache Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/me/cache": {
      "delete": {
        "summary": "Clear Own Cache",
        "description": "Explicit alias of ``DELETE /api/cache`` (stage 9, §66).\n\nSame handler shape, same per-user scope, same ``confirm=true`` gate —\nthe path only says what the code already does: delete the CALLER's\ncache, never anyone else's. Kept side by side with ``/api/cache`` so\nexisting desktop clients keep working while new clients can use the\nunambiguous name.",
        "operationId": "clear_own_cache_api_me_cache_delete",
        "parameters": [
          {
            "name": "confirm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Confirm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Clear Own Cache Api Me Cache Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/feedback/tickets/{ticket_id}": {
      "get": {
        "tags": ["feedback"],
        "summary": "Get My Ticket",
        "description": "One own ticket with its whole conversation, or 404.",
        "operationId": "get_my_ticket_api_feedback_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Ticket Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TicketDetailOut" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/feedback/tickets": {
      "get": {
        "tags": ["admin"],
        "summary": "List Tickets",
        "description": "Every ticket, filtered by status/category/search, paginated.\n\n``q`` matches a subject or a message body as a substring (case-insensitive\nthrough ``ILIKE``; SQLite's ``like`` is case-insensitive for ASCII, which is\nwhat the tests assert). The filters are ANDed and the ordering is the same\n\"newest activity first\" as everywhere else, so the list stays stable while\nsomebody works on it.",
        "operationId": "list_tickets_api_admin_feedback_tickets_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Status"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Category"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "maxLength": 200 },
                { "type": "null" }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminTicketListOut" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminTicketListOut": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/AdminTicketOut" },
            "type": "array",
            "title": "Items"
          },
          "total": { "type": "integer", "title": "Total" },
          "limit": { "type": "integer", "title": "Limit" },
          "offset": { "type": "integer", "title": "Offset" }
        },
        "type": "object",
        "required": ["total", "limit", "offset"],
        "title": "AdminTicketListOut",
        "description": "A page of the admin list plus the total for the pagination control."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncResult": {
        "properties": {
          "ok": { "type": "boolean", "title": "Ok" },
          "last_sync": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Last Sync"
          },
          "courses": { "type": "integer", "title": "Courses", "default": 0 },
          "assignments": {
            "type": "integer",
            "title": "Assignments",
            "default": 0
          },
          "error": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Error"
          },
          "queued": { "type": "boolean", "title": "Queued", "default": false },
          "status": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Status"
          },
          "restarted": {
            "type": "boolean",
            "title": "Restarted",
            "default": false
          }
        },
        "type": "object",
        "required": ["ok"],
        "title": "SyncResult"
      },
...
      "TicketDetailOut": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "category": { "type": "string", "title": "Category" },
          "subject": { "type": "string", "title": "Subject" },
          "status": { "type": "string", "title": "Status" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "messages": {
            "items": { "$ref": "#/components/schemas/MessageOut" },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "subject",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "TicketDetailOut",
        "description": "One ticket with its whole conversation, as its OWNER sees it."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}